// src/_common/enums/ResponseMessages.enum.ts
export enum ResponseMessages {
  SUCCESS = 'SUCCESS',
  BAD_GATEWAY = 'BAD_GATEWAY',
  FORBIDDEN = 'Erişim engellendi',
  UNAUTHORIZED = 'Kimlik doğrulama başarısız',
  BAD_REQUEST = 'Geçersiz istek',
  NOT_FOUND = 'Kaynak bulunamadı',
  INTERNAL_SERVER_ERROR = 'Sunucu hatası',
  VALIDATION_FAILED = 'Doğrulama hatası oluştu',
  DATABASE_ERROR = 'Veritabanı hatası meydana geldi',
  USER_NOT_FOUND = 'Kullanıcı bulunamadı',
  EMAIL_ALREADY_EXISTS = 'E-posta adresi zaten kayıtlı',
  PASSWORD_OR_EMAIL_INCORRECT = 'Şifre veya e-posta adresi yanlış.',
  USER_TYPE_NOT_VALID_FOR_REGISTER = 'USER_TYPE_NOT_VALID_FOR_REGISTER',
  PHONE_ALREADY_EXISTS = 'Telefon numarası zaten kayıtlı',
  EMAIL_NOT_CONFIRMED = 'E-POSTA DOĞRULANMAMIŞ',
  EMAIL_ALREADY_CONFIRMED = 'E-POSTA ZATEN DOĞRULANMIŞ',
  INVALID_CONFIRM_CODE = 'GEÇERSİZ DOĞRULAMA KODU',
  CONFIRM_CODE_EXPIRED = 'DOĞRULAMA KODUNUN SÜRESİ DOLMUŞ',
  EMAIL_CONFIRMED_SUCCESSFULLY = 'E-POSTA BAŞARIYLA DOĞRULANDI',
  CONFIRM_CODE_EXPIRED_NEW_SENT = 'Kodun süresi dolmuş yeni kod gönderildi.',
}